requirements.txt
pandas
numpy
sklearn
scipy
pydeck
streamlit
seaborn
surprise
Pillow
wordcloud